// src/app/models/member.ts
export class Member {
    id: string = "";
    nombre: string = "";
    especialidad: string = "";
    universidad: string = "";
    ano: string = "";
    website: string = "";
    email: string = "";
    direccion: string = "";
    estado: string = "";
    ciudad: string = "";
    telefonos: string = "";
    facebook: string = "";
    instagram: string = "";
    twitter: string = "";
    linkedin: string = "";
    image: string = "";
    created_at: string = "";
    user_id: string = "";

    constructor(id: string,nombre: string,especialidad: string,universidad: string,ano: string,website: string, 
        email: string,direccion: string,estado: string,ciudad: string ,telefonos: string,facebook: string ,instagram: string,
        twitter: string,linkedin: string,image: string,created_at: string,user_id: string ) {}

}